Move the @synchronized decorator to iutil
Move the @synchronized decorator to iutil
so that is can be used in other modules than just
install_tasks.
