+++ pyanaconda/iutil.py
@@ -36,6 +36,7 @@ import sys
 import imp
 import types
 import inspect
+import functools
 
 import requests
 from requests_file import FileAdapter
@@ -1563,3 +1564,22 @@ def item_counter(item_count):
     while index <= item_count:
         yield "%d/%d" % (index, item_count)
         index += 1
+
+def synchronized(wrapped):
+    """A locking decorator for methods.
+
+    The decorator is only intended for methods and the class providing
+    the method also needs to have a Lock/RLock instantiated in self._lock.
+
+    The decorator prevents the wrapped method from being executed until
+    self._lock can be acquired. Once available, it acquires the lock and
+    prevents other decorated methods & other users of self._lock from
+    executing until the wrapped method finishes running.
+    """
+
+    @functools.wraps(wrapped)
+    def _wrapper(self, *args, **kwargs):
+        with self._lock:
+            return wrapped(self, *args, **kwargs)
+    return _wrapper
+
