+++ tests/pyanaconda_tests/iutil_test.py
@@ -17,11 +17,13 @@
 # Red Hat, Inc.
 
 from pyanaconda import iutil
+from pyanaconda.iutil import synchronized
 import unittest
 import os
 import tempfile
 import signal
 import shutil
+from threading import Lock
 from .test_constants import ANACONDA_TEST_DIR
 
 from timer import timer
@@ -840,3 +842,43 @@ class MiscTests(unittest.TestCase):
         counter = iutil.item_counter(-1)
         with self.assertRaises(ValueError):
             next(counter)
+
+    def synchronized_decorator_test(self):
+        """Check that the @synchronized decorator works correctly."""
+
+        # The @synchronized decorator work on methods of classes
+        # that provide self._lock with Lock or RLock instance.
+        class LockableClass(object):
+            def __init__(self):
+                self._lock = Lock()
+
+            def test_method(self):
+                lock_state = self._lock.locked()
+                return lock_state
+
+            @synchronized
+            def sync_test_method(self):
+                lock_state = self._lock.locked()
+                return lock_state
+
+        lockable = LockableClass()
+        self.assertFalse(lockable.test_method())
+        self.assertTrue(lockable.sync_test_method())
+
+        # The @synchronized decorator does not work on classes without self._lock.
+        class NotLockableClass(object):
+            @synchronized
+            def sync_test_method(self):
+                return "Hello world!"
+
+        not_lockable = NotLockableClass()
+        with self.assertRaises(AttributeError):
+            not_lockable.sync_test_method()
+
+        # It also does not work on functions.
+        @synchronized
+        def test_function():
+            return "Hello world!"
+
+        with self.assertRaises(TypeError):
+            test_function()
